Extract a transform() library for reuse.
Now most of main.js is just test harness.
